# bcm/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Business Capability Model API",
        "version": "0.1.0"
    },
    "paths": {
        "/users": {
            "get": {
                "summary": "Get Active Users",
                "description": "Get all active users and their locked capabilities.",
                "operationId": "get_active_users_users_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/UserSession"
                                    },
                                    "type": "array",
                                    "title": "Response Get Active Users Users Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create User Session",
                "description": "Create a new user session.",
                "operationId": "create_user_session_users_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSession"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{session_id}": {
            "delete": {
                "summary": "Remove User Session",
                "description": "Remove a user session.",
                "operationId": "remove_user_session_users__session_id__delete",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/capabilities/lock/{capability_id}": {
            "post": {
                "summary": "Lock Capability",
                "description": "Lock a capability for editing.",
                "operationId": "lock_capability_capabilities_lock__capability_id__post",
                "parameters": [
                    {
                        "name": "capability_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Capability Id"
                        }
                    },
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/capabilities/unlock/{capability_id}": {
            "post": {
                "summary": "Unlock Capability",
                "description": "Unlock a capability.",
                "operationId": "unlock_capability_capabilities_unlock__capability_id__post",
                "parameters": [
                    {
                        "name": "capability_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Capability Id"
                        }
                    },
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/capabilities": {
            "post": {
                "summary": "Create Capability",
                "description": "Create a new capability.",
                "operationId": "create_capability_capabilities_post",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CapabilityCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Create Capability Capabilities Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Get Capabilities",
                "description": "Get capabilities, optionally filtered by parent_id.\nIf hierarchical=True, returns full tree structure under the parent_id.\nIf hierarchical=False, returns flat list of immediate children.",
                "operationId": "get_capabilities_capabilities_get",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Parent Id"
                        }
                    },
                    {
                        "name": "hierarchical",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Hierarchical"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    },
                                    "title": "Response Get Capabilities Capabilities Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/capabilities/{capability_id}": {
            "get": {
                "summary": "Get Capability",
                "description": "Get a capability by ID.",
                "operationId": "get_capability_capabilities__capability_id__get",
                "parameters": [
                    {
                        "name": "capability_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Capability Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Get Capability Capabilities  Capability Id  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Capability",
                "description": "Update a capability.",
                "operationId": "update_capability_capabilities__capability_id__put",
                "parameters": [
                    {
                        "name": "capability_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Capability Id"
                        }
                    },
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CapabilityUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Update Capability Capabilities  Capability Id  Put"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Capability",
                "description": "Delete a capability.",
                "operationId": "delete_capability_capabilities__capability_id__delete",
                "parameters": [
                    {
                        "name": "capability_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Capability Id"
                        }
                    },
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/capabilities/{capability_id}/context": {
            "get": {
                "summary": "Get Capability Context",
                "description": "Get a capability's context (parent and children).",
                "operationId": "get_capability_context_capabilities__capability_id__context_get",
                "parameters": [
                    {
                        "name": "capability_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Capability Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Get Capability Context Capabilities  Capability Id  Context Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "User": {
                "properties": {
                    "nickname": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1,
                        "title": "Nickname"
                    }
                },
                "type": "object",
                "required": [
                    "nickname"
                ],
                "title": "User"
            },
            "UserSession": {
                "properties": {
                    "session_id": {
                        "type": "string",
                        "title": "Session Id"
                    },
                    "nickname": {
                        "type": "string",
                        "title": "Nickname"
                    },
                    "locked_capabilities": {
                        "items": {
                            "type": "integer"
                        },
                        "type": "array",
                        "title": "Locked Capabilities"
                    }
                },
                "type": "object",
                "required": [
                    "session_id",
                    "nickname"
                ],
                "title": "UserSession"
            },
...
        }
    }
}